+++ lib/mwlib/tests/MW/Logger/FileTest.php
@@ -1,41 +1,35 @@
 <?php
 
-namespace Aimeos\MW\Logger;
-
-
 /**
  * @license LGPLv3, http://opensource.org/licenses/LGPL-3.0
  * @copyright Metaways Infosystems GmbH, 2013
  * @copyright Aimeos (aimeos.org), 2015-2016
  */
+
+
+namespace Aimeos\MW\Logger;
+
+
 class FileTest extends \PHPUnit\Framework\TestCase
 {
 	private $object;
 	private $filename;
 
 
-	/**
-	 * Sets up the fixture, for example, opens a network connection.
-	 * This method is called before a test is executed.
-	 *
-	 * @access protected
-	 */
 	protected function setUp()
 	{
 		$this->filename = 'loggertest.log';
 		$this->object = new \Aimeos\MW\Logger\File( $this->filename );
 	}
 
-	/**
-	 * Tears down the fixture, for example, closes a network connection.
-	 * This method is called after a test is executed.
-	 *
-	 * @access protected
-	 */
+
 	protected function tearDown()
 	{
-		unlink( 'loggertest.log' );
+		if( file_exists( 'loggertest.log' ) ) {
+			unlink( 'loggertest.log' );
+		}
 	}
+
 
 	public function testLog()
 	{
@@ -63,6 +57,7 @@ class FileTest extends \PHPUnit\Framework\TestCase
 		$this->object->log( 'wrong log level', -1);
 	}
 
+
 	public function testScalarLog()
 	{
 		$this->object->log( array ( 'scalar', 'errortest' ) );
@@ -84,6 +79,7 @@ class FileTest extends \PHPUnit\Framework\TestCase
 		$this->assertEquals( \Aimeos\MW\Logger\Base::ERR, $msg[3] );
 		$this->assertEquals( '["scalar","errortest"]', $msg[4] );
 	}
+
 
 	public function testLogCrit()
 	{
@@ -107,21 +103,14 @@ class FileTest extends \PHPUnit\Framework\TestCase
 		$this->assertEquals( 'critical', $msg[4] );
 	}
 
+
 	public function testLogWarn()
 	{
 		$this->object->log( 'debug', \Aimeos\MW\Logger\Base::WARN );
 
-		if( !file_exists( $this->filename ) ) {
-			throw new \RuntimeException( 'No test file found' );
-		}
-
-		$lines = explode( PHP_EOL, file_get_contents( $this->filename ) );
-		$msg = $lines[0];
-
-		if( $msg !== '' ) {
-			throw new \RuntimeException( 'Log record found but none expected' );
-		}
+		$this->assertFalse( file_exists( $this->filename ) );
 	}
+
 
 	public function testFacility()
 	{
@@ -141,11 +130,12 @@ class FileTest extends \PHPUnit\Framework\TestCase
 		$this->assertEquals( '<auth>', $msg[0] );
 	}
 
+
 	public function testFacilityLimited()
 	{
 		$this->object = new \Aimeos\MW\Logger\File( $this->filename, \Aimeos\MW\Logger\Base::ERR, array( 'test' ) );
 		$this->object->log( 'user auth', \Aimeos\MW\Logger\Base::ERR, 'auth' );
 
-		$this->assertEquals( '', file_get_contents( $this->filename ) );
+		$this->assertFalse( file_exists( $this->filename ) );
 	}
 }
